requirements.txt
future==0.16.0
-e sdk/python
nose==1.3.7
rednose==1.2.2
pyserial==3.4
